# Graphics buffer performance.