--- a/Elements/src/Geometry/GraphicsBuffers.cs
+++ b/Elements/src/Geometry/GraphicsBuffers.cs
@@ -19,6 +19,11 @@ namespace Elements.Geometry
         void AddVertex(Vector3 position, Vector3 normal, UV uv, Color? color = null);
 
         /// <summary>
+        /// Add vertices to the graphics buffers
+        /// </summary>
+        void AddVertices(IList<(Vector3 position, Vector3 normal, UV uv, Color color)> vertices);
+
+        /// <summary>
         /// Add a vertex to the graphics buffers.
         /// </summary>
         /// <param name="x"></param>
@@ -225,6 +230,78 @@ namespace Elements.Geometry
         }
 
         /// <summary>
+        /// Add vertices to the graphics buffers.
+        /// </summary>
+        public void AddVertices(IList<(Vector3 position, Vector3 normal, UV uv, Color color)> vertices)
+        {
+            var vertexStep = sizeof(float) * 3;
+            var normalStep = sizeof(float) * 3;
+            var uvStep = sizeof(float) * 2;
+            var colorStep = sizeof(float) * 3;
+
+            var allPositions = new byte[vertices.Count * vertexStep];
+            var allNormals = new byte[vertices.Count * normalStep];
+            var allUvs = new byte[vertices.Count * uvStep];
+            var allColors = new byte[vertices.Count * colorStep];
+
+            var hasVertexColors = false;
+
+            for (var i = 0; i < vertices.Count; i++)
+            {
+                var (position, normal, uv, color) = vertices[i];
+
+                Buffer.BlockCopy(BitConverter.GetBytes((float)position.X), 0, allPositions, vertexStep * i, sizeof(float));
+                Buffer.BlockCopy(BitConverter.GetBytes((float)position.Y), 0, allPositions, vertexStep * i + sizeof(float), sizeof(float));
+                Buffer.BlockCopy(BitConverter.GetBytes((float)position.Z), 0, allPositions, vertexStep * i + sizeof(float) * 2, sizeof(float));
+                this.VMax[0] = Math.Max(this.VMax[0], position.X);
+                this.VMax[1] = Math.Max(this.VMax[1], position.Y);
+                this.VMax[2] = Math.Max(this.VMax[2], position.Z);
+                this.VMin[0] = Math.Min(this.VMin[0], position.X);
+                this.VMin[1] = Math.Min(this.VMin[1], position.Y);
+                this.VMin[2] = Math.Min(this.VMin[2], position.Z);
+
+                Buffer.BlockCopy(BitConverter.GetBytes((float)normal.X), 0, allNormals, normalStep * i, sizeof(float));
+                Buffer.BlockCopy(BitConverter.GetBytes((float)normal.Y), 0, allNormals, normalStep * i + sizeof(float), sizeof(float));
+                Buffer.BlockCopy(BitConverter.GetBytes((float)normal.Z), 0, allNormals, normalStep * i + sizeof(float) * 2, sizeof(float));
+                this.NMax[0] = Math.Max(this.NMax[0], normal.X);
+                this.NMax[1] = Math.Max(this.NMax[1], normal.Y);
+                this.NMax[2] = Math.Max(this.NMax[2], normal.Z);
+                this.NMin[0] = Math.Min(this.NMin[0], normal.X);
+                this.NMin[1] = Math.Min(this.NMin[1], normal.Y);
+                this.NMin[2] = Math.Min(this.NMin[2], normal.Z);
+
+                Buffer.BlockCopy(BitConverter.GetBytes((float)uv.U), 0, allUvs, uvStep * i, sizeof(float));
+                Buffer.BlockCopy(BitConverter.GetBytes((float)uv.V), 0, allUvs, uvStep * i + sizeof(float), sizeof(float));
+                this.UVMax[0] = Math.Max(this.UVMax[0], uv.U);
+                this.UVMax[1] = Math.Max(this.UVMax[1], uv.V);
+                this.UVMin[0] = Math.Min(this.UVMin[0], uv.U);
+                this.UVMin[1] = Math.Min(this.UVMin[1], uv.V);
+
+                if (color != default)
+                {
+                    hasVertexColors = true;
+                    Buffer.BlockCopy(BitConverter.GetBytes((float)color.Red), 0, allColors, colorStep * i, sizeof(float));
+                    Buffer.BlockCopy(BitConverter.GetBytes((float)color.Green), 0, allColors, colorStep * i + sizeof(float), sizeof(float));
+                    Buffer.BlockCopy(BitConverter.GetBytes((float)color.Blue), 0, allColors, colorStep * i + sizeof(float) * 2, sizeof(float));
+                    this.CMax[0] = Math.Max(this.CMax[0], color.Red);
+                    this.CMax[1] = Math.Max(this.CMax[1], color.Green);
+                    this.CMax[2] = Math.Max(this.CMax[2], color.Blue);
+                    this.CMin[0] = Math.Min(this.CMin[0], color.Red);
+                    this.CMin[1] = Math.Min(this.CMin[1], color.Green);
+                    this.CMin[2] = Math.Min(this.CMin[2], color.Blue);
+                }
+            }
+
+            this.Vertices.AddRange(allPositions);
+            this.Normals.AddRange(allNormals);
+            this.UVs.AddRange(allUvs);
+            if (hasVertexColors)
+            {
+                this.Colors.AddRange(allColors);
+            }
+        }
+
+        /// <summary>
         /// Add an index to the graphics buffers.
         /// </summary>
         /// <param name="index">The index to add.</param>
@@ -261,5 +338,6 @@ namespace Elements.Geometry
             var intSize = sizeof(int);
             return (floatSize * 3 + floatSize * 3 + floatSize * 4 + intSize + floatSize * 2) * _preallocationVertexCount;
         }
+
     }
 }