BACKGROUND:
- Web assembly performance is impacted by large copy operations like those when we convert a `List<byte>` to `byte[]` using `buffer.ToArray()`.

DESCRIPTION:
- Preallocate buffers in `GraphicsBuffers`
- Remove numerous uses of `ToArray` to reduce copying.

REQUIRED:
- [ ] All changes are up to date in `CHANGELOG.md`. **n/a**

Before:
|                                        Method |     Mean |   Error |  StdDev |      Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|---------------------------------------------- |---------:|--------:|--------:|-----------:|----------:|----------:|----------:|
| 'Create all HSS beams and serialize to glTF.' | 146.2 ms | 2.51 ms | 4.58 ms | 26000.0000 | 2333.3333 | 1333.3333 |     97 MB |

After:
|                                        Method |     Mean |    Error |   StdDev |      Gen 0 |     Gen 1 | Allocated |
|---------------------------------------------- |---------:|---------:|---------:|-----------:|----------:|----------:|
| 'Create all HSS beams and serialize to glTF.' | 85.31 ms | 1.402 ms | 1.311 ms | 22000.0000 | 3000.0000 |     80 MB |

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/hypar-io/elements/845)
<!-- Reviewable:end -->